requirements.txt
Pillow>=10.0.0
scikit-image>=0.21.0
PyPDF2>=3.0.0
imagehash>=4.3.1
numpy>=1.24.0

